refactor(util): split SocketUtil.h into focused sub-headers
Split the monolithic SocketUtil.h (2355 lines) into 14 focused sub-headers
in include/core/SocketUtil/ for improved maintainability and reduced
compilation dependencies.

New sub-headers:
- Core.h: Bit macros, MIN/MAX, ARRAY_LENGTH
- Constants.h: Port limits, buffer sizes
- Hash.h: Golden ratio, DJB2, prime31 hash functions
- Time.h: Monotonic clock, timespec conversions
- Timeout.h: Deadline calculations
- TTL.h: TTL expiry utilities
- ByteOrder.h: Big-endian pack/unpack
- StringUtils.h: Arena strdup, safe copy
- URL.h: Percent-decoding
- Arithmetic.h: Safe overflow-checked arithmetic
- IO.h: EINTR-safe read/write
- Error.h: Error formatting macros
- Exception.h: Exception infrastructure
- MutexArena.h: Mutex + arena allocation pattern

The master SocketUtil.h now includes all sub-headers for backward
compatibility. Existing code that includes SocketUtil.h continues to work
unchanged.

Fixes #2556
